// app/(auth-pages)/schema.ts
import z from "zod";

// login schema

export const loginSchema = z.object({
  email: z.string().email({ message: "Invalid email address" }),
  password: z.string({ message: "Password is required" }).min(6, { message: 'Password must be at least 6 characters'}),
});


// register schema

export const registerSchema = z.object({
    name: z.string().min(1, { message: "Name is required" }),
    email: z.string().email({ message: "Invalid email address" }),
    password: z.string().min(6, { message: "Password must be at least 6 characters" }),
  });